pyproject.toml
[tool.poetry]
name = "restaurant_backend"
version = "0.1.0"
description = ""
authors = ["Ankush <ankush357159@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
